refactor(cli): extract init command body into framework-agnostic initAction
Moves ~1100 lines of logic out of the oclif InitCommand into
actions/init/initAction.ts, leaving init.ts as a thin wrapper that
maps oclif flags → InitOptions and catches InitError.

Flow files (initApp, initStudio, initNextJs, scaffoldTemplate) now
accept `options: InitOptions` instead of individual flag params,
and throw InitError directly instead of receiving an `error` callback.

Zero test assertion changes — all 14 init test files pass as-is.
