+++ packages/@sanity/cli/src/actions/init/initStudio.ts
@@ -9,9 +9,11 @@ import {updateProjectInitializedAt} from '../../services/projects.js'
 import {type InitStepResult} from '../../telemetry/init.telemetry.js'
 import {type PackageManager} from '../../util/packageManager/packageManagerChoice.js'
 import {type EditorName} from '../mcp/editorConfigs.js'
+import {InitError} from './initError.js'
 import {getPostInitMCPPrompt} from './initHelpers.js'
 import {type RepoInfo} from './remoteTemplate.js'
 import {scaffoldAndInstall, selectTemplate} from './scaffoldTemplate.js'
+import {type InitOptions} from './types.js'
 
 const debug = subdebug('init')
 
@@ -23,70 +25,49 @@ async function promptForDatasetImport(message?: string): Promise<boolean> {
 }
 
 export async function initStudio({
-  autoUpdates,
   datasetName,
   defaults,
   displayName,
-  error,
-  git,
-  importDataset,
   isFirstProject,
   mcpConfigured,
-  noGit,
+  options,
   organizationId,
   output,
   outputPath,
-  overwriteFiles,
-  packageManager,
   projectId,
   remoteTemplateInfo,
   sluggedName,
-  template,
-  templateToken,
   trace,
-  typescript,
-  unattended,
   workDir,
 }: {
-  autoUpdates: boolean
   datasetName: string
   defaults: {projectName: string}
   displayName: string
-  error: Output['error']
-  git?: boolean | string
-  importDataset?: boolean
   isFirstProject: boolean
   mcpConfigured: EditorName[]
-  noGit?: boolean
+  options: InitOptions
   organizationId: string | undefined
   output: Output
   outputPath: string
-  overwriteFiles?: boolean
-  packageManager?: string
   projectId: string
   remoteTemplateInfo: RepoInfo | undefined
   sluggedName: string
-  template?: string
-  templateToken?: string
   trace: TelemetryTrace<TelemetryUserProperties, InitStepResult>
-  typescript?: boolean
-  unattended: boolean
   workDir: string
 }): Promise<void> {
+  const {importDataset, unattended} = options
   const {
     template: resolvedTemplate,
     templateName,
     useTypeScript,
   } = await selectTemplate({
+    options,
     remoteTemplateInfo,
-    template,
     trace,
-    typescript,
-    unattended,
   })
 
   if (!remoteTemplateInfo && !resolvedTemplate) {
-    error(`Template "${templateName}" not found`, {exit: 1})
+    throw new InitError(`Template "${templateName}" not found`, 1)
   }
 
   // If the template has a sample dataset, prompt the user whether or not we should import it
@@ -108,24 +89,18 @@ export async function initStudio({
   }
 
   const {pkgManager} = await scaffoldAndInstall({
-    autoUpdates,
     datasetName,
     defaults,
     displayName,
-    git,
-    noGit,
+    options,
     organizationId,
     output,
     outputPath,
-    overwriteFiles,
-    packageManager,
     projectId,
     remoteTemplateInfo,
     sluggedName,
     templateName,
-    templateToken,
     trace,
-    unattended,
     useTypeScript,
     workDir,
   })
@@ -134,7 +109,7 @@ export async function initStudio({
   if (shouldImport && resolvedTemplate?.datasetUrl) {
     const token = await getCliToken()
     if (!token) {
-      return error('Authentication required to import dataset', {exit: 1})
+      throw new InitError('Authentication required to import dataset', 1)
     }
     await ImportDatasetCommand.run(
       [
